requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6
openai==1.93.3
pandas==2.3.1
python-dotenv==1.1.1
